Lab4

+++ Topdown2DGame/Assets/Scripts/Enemy/EnemyBehaviour.cs
@@ -17,7 +17,7 @@ public class EnemyBehaviour
         isDead = false;
     }
 
-    public virtual void TakeDamage(int amount) 
+    public virtual void TakeDamage(float amount) 
     {
